// src/shared/types/TelegramCallbackData.types.ts
export enum TelegramCallbackDataEnum {
    SETTINGS_NEW_USER = 'settings_new_user',
    ADD_TRADING_PAIR = 'add_trading_pair',
    CHOOSE_TRADING_PAIR = 'choose_trading_pair_',
    DELETE_TRADING_PAIR = 'delete_trading_pair_',
    LAST_SETTINGS_STEP = 'last_settings_step',
    CHANGE_SIGNAL_NOTIFICATION = 'change_signal_notification',
    SETTINGS = 'settings',
    MAIN_MENU = 'main_menu',
}